Add additional initializer for `Font`
This would allow using a custom monospace font for highlighting.
This is especially useful if, say, the code is embedded into some sort of presentation 😊.
An assert could be added to make sure the font is monospace,
but I feel like that's not required.

If somebody wants to display their code in Comic Sans, all the more power to them